+++ src/index.js
@@ -3,19 +3,19 @@
 
 import Accordion from './Accordion';
 import Block from './Block';
-import Button from './Button';
+import Button from './atomic/atoms/Button';
 import Card from './Card';
 import Checkbox from './Checkbox';
 import DeckSwiper from './DeckSwiper';
-import Icon from './Icon';
-import Input from './Input';
+import Icon from './atomic/ions/Icon';
+import Input from './atomic/atoms/Input';
 import NavBar from './NavBar';
 import Radio from './Radio';
 import Slider from './Slider';
 import Switch from './Switch';
-import Text from './Text';
+import Text from './atomic/ions/Text';
 import Toast from './Toast';
-import Link from './Link';
+import Link from './atomic/atoms/Link';
 
 
 import theme, { withGalio, GalioProvider } from './theme';
