+++ src/theme/sizes.js
@@ -58,6 +58,7 @@ const SIZES = {
   INPUT_VERTICAL_TEXT: 14,
   INPUT_VERTICAL_LABEL: BASE / 2,
   INPUT_TEXT: BASE * 0.875,
+  INPUT_ROUNDED: BASE * 1.5,
 
   // NavBar styles
   NAVBAR_HEIGHT: BASE * 4.125,
